refactor: NotePreimage --> Note
