+++ yarn-project/aztec.js/src/pxe_client.ts
@@ -17,7 +17,7 @@ import {
   L2BlockL2Logs,
   L2Tx,
   LogId,
-  NotePreimage,
+  Note,
   PXE,
   Tx,
   TxExecutionRequest,
@@ -43,7 +43,7 @@ export const createPXEClient = (url: string, fetch = makeFetch([1, 2, 3], true))
       Point,
       Fr,
       GrumpkinScalar,
-      NotePreimage,
+      Note,
       ExtendedNote,
       AuthWitness,
       L2Tx,
