+++ yarn-project/pxe/src/pxe_http/pxe_http_server.ts
@@ -13,7 +13,7 @@ import {
   L2BlockL2Logs,
   L2Tx,
   LogId,
-  NotePreimage,
+  Note,
   PXE,
   Tx,
   TxExecutionRequest,
@@ -48,7 +48,7 @@ export function createPXERpcServer(pxeService: PXE): JsonRpcServer {
       Point,
       Fr,
       GrumpkinScalar,
-      NotePreimage,
+      Note,
       ExtendedNote,
       AuthWitness,
       L2Block,
