+++ yarn-project/types/src/logs/l1_note_payload/index.ts
@@ -1,3 +1,3 @@
 export * from './encrypt_buffer.js';
-export * from './note_preimage.js';
+export * from './note.js';
 export * from './l1_note_payload.js';
